python/requirements.txt
dash==2.8.1
twilio
numpy
pandas
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
dash_bootstrap_components
